scripts/requirements.txt
pytube==15.0.0
requests==2.31.0
Pillow==10.2.0